Use native Go calls at internal adapter boundaries

+++ src/legacy/briefing_window.go
@@ -1,11 +1,5 @@
 package legacy
 
-/*
-#include "defs.h"
-
-*/
-import "C"
-
 import (
 	"image"
 	"math"
@@ -53,7 +47,7 @@ func briefingCreateWindow() *gui.Window {
 	*memmap.PtrUint32(0x5D4594, 831284) = uint32(ox)
 	*memmap.PtrUint32(0x5D4594, 831288) = uint32(oy)
 	parent := GetClient().Cli().GUI.NewWindowRaw(nil, 56, 0, 0, int(nox_win_width), int(nox_win_height), briefingEvent(briefingBackgroundEvent))
-	legacyGlobals.dword_5d4594_831236 = C.uint32_t(uint32(uintptr(parent.C())))
+	legacyGlobals.dword_5d4594_831236 = uint32(uintptr(parent.C()))
 	if parent == nil {
 		return nil
 	}
