+++ src/legacy/server_browser_events.go
@@ -1,13 +1,5 @@
 package legacy
 
-/*
-#include "GAME1_2.h"
-#include "GAME2_2.h"
-#include "GAME2_3.h"
-#include "client__shell__noxworld.h"
-*/
-import "C"
-
 import (
 	"encoding/binary"
 	"github.com/opennox/opennox/v1/client/gui"
@@ -92,10 +84,10 @@ func browserEvent(w *gui.Window, ev gui.WindowEvent) gui.WindowEventResp {
 		Sub_4373A0()
 	case 10047, 10048, 10049, 10050, 10051:
 		browserListReset()
-		nox_wol_servers_sortBtnHandler_4A0290(C.int(int32(id)))
+		nox_wol_servers_sortBtnHandler_4A0290(int32(id))
 		browserListResort()
 	case 10054, 10055, 10056, 10057:
-		browserUI.region = C.int(int32(id - 10054))
+		browserUI.region = int32(id - 10054)
 		browserShowRegion()
 		Nox_client_refreshServerList_4378B0()
 		x, y := uint32(408), uint32(239)
@@ -164,7 +156,7 @@ func browserEvent(w *gui.Window, ev gui.WindowEvent) gui.WindowEventResp {
 			}
 			SendXXX_5550D0(netip.AddrPortFrom(int2ip(uint32(nox_client_getServerAddr_43B300())), uint16(nox_client_getServerPort_43B320())), buf[:])
 			browserUI.connectionState = 3
-			browserUI.connectionDeadline = C.uint64_t(uint64(uint32(PlatformTicks()) + 20000))
+			browserUI.connectionDeadline = uint64(uint32(PlatformTicks()) + 20000)
 			Sub_449EA0(0)
 		case 10:
 			Sub_449E60(4)
@@ -176,7 +168,7 @@ func browserEvent(w *gui.Window, ev gui.WindowEvent) gui.WindowEventResp {
 			nox_game_showGameSel_4379F0()
 		default:
 			if browserUI.retry != 0 {
-				browserUI.refreshDeadline = C.uint64_t(uint64(uint32(PlatformTicks()) + 1000))
+				browserUI.refreshDeadline = uint64(uint32(PlatformTicks()) + 1000)
 			}
 		}
 	case 4002:
